remove warnings, dual and verbose tests

+++ text_renderer.cpp
@@ -17,15 +17,15 @@ struct GeometrySink : public ID2D1GeometrySink {
 
 	D2D1_POINT_2F previous;
 	void AddBeziers(const D2D1_BEZIER_SEGMENT *points, UINT32 count) override {
-		for (int i = 0; i < count; ++i)
+		for (unsigned i = 0; i < count; ++i)
 			cubicSplines.push_back({ previous, points[i].point1, points[i].point2, previous = points[i].point3 });
 	}
 	void AddQuadraticBeziers(const D2D1_QUADRATIC_BEZIER_SEGMENT* points, UINT32 count) override {
-		for (int i = 0; i < count; ++i)
+		for (unsigned i = 0; i < count; ++i)
 			quadraticSplines.push_back({ previous, points[i].point1, previous = points[i].point2 });
 	}
 	void AddLines(const D2D1_POINT_2F *points, UINT32 count) override {
-		for (int i = 0; i < count; ++i)
+		for (unsigned i = 0; i < count; ++i)
 			lines.push_back({ previous, previous = points[i] });
 	}
 	void AddArc(const D2D1_ARC_SEGMENT* arc) override { printf("arcs not implemented!\n"); }
@@ -85,7 +85,7 @@ HRESULT TextRenderer::DrawGlyphRun(void*, FLOAT baseX, FLOAT baseY, DWRITE_MEASU
 
 	HRESULT result = factory->TranslateColorGlyphRun(D2D1_POINT_2F{ baseX, baseY }, run, runDesc, DWRITE_GLYPH_IMAGE_FORMATS_TRUETYPE | DWRITE_GLYPH_IMAGE_FORMATS_COLR | DWRITE_GLYPH_IMAGE_FORMATS_CFF, mode, nullptr, 0, &enumerator);
 	if (result == DWRITE_E_NOCOLOR) {
-		for (int i = 0; i < run->glyphCount; ++i) {
+		for (unsigned i = 0; i < run->glyphCount; ++i) {
 			UINT16 glyph = run->glyphIndices[i];
 			auto glyphIterator = glyphToOffset.find(std::make_pair(run->fontEmSize, glyph));
 			std::pair<unsigned, unsigned> range;
@@ -111,14 +111,14 @@ HRESULT TextRenderer::DrawGlyphRun(void*, FLOAT baseX, FLOAT baseY, DWRITE_MEASU
 
 				if (inds[3] > inds[0]) {
 					low = high = points[inds[0]];
-					for (int j = inds[0] + 1; j < inds[3]; ++j) {
+					for (unsigned j = inds[0] + 1; j < inds[3]; ++j) {
 						D2D1_POINT_2F p = points[j];
 						if (p.x < low.x) low.x = p.x;
 						if (p.y < low.y) low.y = p.y;
 						if (p.x > high.x) high.x = p.x;
 						if (p.y > high.y) high.y = p.y;
 					}
-					for (int j = inds[0]; j < inds[3]; ++j) {
+					for (unsigned j = inds[0]; j < inds[3]; ++j) {
 						points[j].x -= low.x;
 						points[j].y -= low.y;
 					}
@@ -152,7 +152,7 @@ HRESULT TextRenderer::DrawGlyphRun(void*, FLOAT baseX, FLOAT baseY, DWRITE_MEASU
 	}
 	else {
 		enumerator->Release();
-		for (int i = 0; i < run->glyphCount; ++i) {
+		for (unsigned i = 0; i < run->glyphCount; ++i) {
 			UINT16 glyph = run->glyphIndices[i];
 			auto glyphIterator = glyphToOffset.find(std::make_pair(run->fontEmSize, glyph));
 
@@ -162,7 +162,7 @@ HRESULT TextRenderer::DrawGlyphRun(void*, FLOAT baseX, FLOAT baseY, DWRITE_MEASU
 
 				glyphIterator = glyphToOffset.insert(glyphIterator, std::make_pair(std::make_pair(run->fontEmSize, glyph), unsigned(colorRanges.size())));
 
-				range.first = colors.size();
+				range.first = (uint32_t)colors.size();
 
 				DWRITE_GLYPH_RUN tmpRun = *run;
 				tmpRun.glyphIndices = &glyph;
@@ -207,17 +207,17 @@ HRESULT TextRenderer::DrawGlyphRun(void*, FLOAT baseX, FLOAT baseY, DWRITE_MEASU
 				enumerator->Release();
 
 				//printf("SNAB\n");
-				range.second = colors.size();
+				range.second = (uint32_t)colors.size();
 
 				low = high = points[pointIndices[range.first][0]];
-				for (int j = pointIndices[range.first][0] + 1; j < pointIndices[range.second - 1][3]; ++j) {
+				for (unsigned j = pointIndices[range.first][0] + 1; j < pointIndices[range.second - 1][3]; ++j) {
 					D2D1_POINT_2F p = points[j];
 					if (p.x < low.x) low.x = p.x;
 					if (p.y < low.y) low.y = p.y;
 					if (p.x > high.x) high.x = p.x;
 					if (p.y > high.y) high.y = p.y;
 				}
-				for (int j = pointIndices[range.first][0]; j < pointIndices[range.second - 1][3]; ++j) {
+				for (unsigned j = pointIndices[range.first][0]; j < pointIndices[range.second - 1][3]; ++j) {
 					points[j].x -= low.x;
 					points[j].y -= low.y;
 				}
@@ -301,7 +301,7 @@ void Font::drawText(const std::wstring& text, float x, float y, float size, floa
 	renderer.factory->CreateTextFormat(fontName.c_str(), nullptr, nullptr, 0, size, L"", &format);
 
 	IDWriteTextLayout4* layout;
-	renderer.factory->CreateTextLayout(text.c_str(), text.length(), format, maxwidth, maxheight, (IDWriteTextLayout**)&layout);
+	renderer.factory->CreateTextLayout(text.c_str(), (UINT32)text.length(), format, maxwidth, maxheight, (IDWriteTextLayout**)&layout);
 
 	layout->Draw(nullptr, &renderer, x, y);
 	size_t count = renderer.updateBuffers();
